More unmerged target issues fixed
