+++ examples/cpp/AutoCalibration/CMakeLists.txt
@@ -1,4 +1,6 @@
 project(auto_calibration_examples)
 cmake_minimum_required(VERSION 3.10)
 
-dai_add_example(auto_calibration_example auto_calibration_example.cpp ON OFF)
+if(DEPTHAI_MERGED_TARGET)
+    dai_add_example(auto_calibration_example auto_calibration_example.cpp ON OFF)
+endif()
