+++ src/pipeline/node/Camera.cpp
@@ -146,20 +146,6 @@ std::shared_ptr<Camera> Camera::build(CameraBoardSocket boardSocket,
     isBuilt = true;
     return std::static_pointer_cast<Camera>(shared_from_this());
 }
-#ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-std::shared_ptr<Camera> Camera::build(CameraBoardSocket boardSocket, ReplayVideo& replay) {
-    auto cam = build(boardSocket);
-    cam->setMockIsp(replay);
-    return cam;
-}
-
-std::shared_ptr<Camera> Camera::build(ReplayVideo& replay) {
-    auto cam = build(CameraBoardSocket::AUTO);
-    cam->setMockIsp(replay);
-    return cam;
-}
-#endif
-
 Camera::Properties& Camera::getProperties() {
     properties.initialControl = initialControl;
     return properties;
@@ -262,38 +248,6 @@ Node::Output* Camera::requestIspOutput(std::optional<float> fps) {
 Node::Output* Camera::requestOutput(const Capability& capability, bool onHost) {
     return pimpl->requestOutput(*this, capability, onHost);
 }
-
-#ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-Camera& Camera::setMockIsp(ReplayVideo& replay) {
-    if(!replay.getReplayVideoFile().empty()) {
-        auto [width, height] = replay.getSize();
-        double fps = (double)replay.getFps();
-        if(width <= 0 || height <= 0) {
-            const auto& [vidWidth, vidHeight, vidFps] = utility::getVideoSize(replay.getReplayVideoFile().string());
-            width = vidWidth;
-            height = vidHeight;
-            fps = vidFps;
-        }
-        properties.mockIspWidth = width;
-        properties.mockIspHeight = height;
-        properties.mockIspFps = fps;
-
-        auto device = getParentPipeline().getDefaultDevice();
-        if(device) {
-            if(device->getPlatform() == Platform::RVC2) {
-                replay.setOutFrameType(ImgFrame::Type::YUV420p);
-            } else {
-                replay.setOutFrameType(ImgFrame::Type::NV12);
-            }
-        }
-
-        replay.out.link(mockIsp);
-    } else {
-        throw std::runtime_error("ReplayVideo video path not set");
-    }
-    return *this;
-}
-#endif
 
 void Camera::buildStage1() {
     return pimpl->buildStage1(*this);
